requirements.txt
PySide6-Fluent-Widgets>=1.5.5
py7zr>=0.20.0
requests>=2.31.0